+++ src/main/java/com/kanzariya/chatsphere/service/MessageService.java
@@ -9,7 +9,7 @@ import com.kanzariya.chatsphere.entity.Message;
 
 
 public interface MessageService {
-	public Message sendMessage(Long senderId, Long receiverId, Long groupId, String content);
+	public Message sendMessage(int senderId, int receiverId, Long groupId, String content);
 	 public List<Message> getMessages(Long receiverId);
 	 public List<Message> getMessagesByGroup(Long groupId);
 	 public Message updateMessage(Long id,String content);
